♻️ refactor: admin category create 구조 정리 및 API URL 슬래시 통일
- admin 폴더 구조 반영
- category create API URL 슬래시 통일
- 테스트 코드 URL 수정

# Conflicts:
#	config/urls.py

# Conflicts:
#	config/urls.py
