+++ apps/qna/admin/category_views.py
@@ -6,11 +6,11 @@ from rest_framework.request import Request
 from rest_framework.response import Response
 from rest_framework.views import APIView
 
+from apps.qna.admin.category_services import CategoryService
 from apps.qna.serializers.category_serializers import (
     AdminCategoryCreateResponseSerializer,
     AdminCategoryCreateSerializer,
 )
-from apps.qna.services.category_services import CategoryService
 
 
 class AdminCategoryCreateAPIView(APIView):
