+++ apps/qna/tests/test_categories/test_admin_category_create.py
@@ -18,7 +18,7 @@ class AdminCategoryCreateAPITest(APITestCase):
             email="testadmin@example.com",
             password="test1234",
         )
-        cls.url = "/api/v1/admin/qna/categories"
+        cls.url = "/api/v1/admin/qna/categories/"
 
     # 테스트마다 실행됨 (상태 객체 초기화)
     def setUp(self) -> None:
