+++ apps/qna/urls/admin_urls.py
@@ -1,11 +1,11 @@
 from django.urls import path
 
-from apps.qna.views.category_views import AdminCategoryCreateAPIView
+from apps.qna.admin.category_views import AdminCategoryCreateAPIView
 
 urlpatterns = [
     # 어드민 카테고리
     path(
-        "categories",
+        "categories/",
         AdminCategoryCreateAPIView.as_view(),
         name="admin-category-create",
     ),
