+++ config/urls.py
@@ -10,6 +10,8 @@ from drf_spectacular.views import (
 urlpatterns: list[URLPattern | URLResolver] = [
     path("api/v1/posts/", include("apps.posts.urls")),
     path("api/v1/", include("apps.qna.urls.urls")),
+    path("api/v1/admin/qna/", include("apps.qna.urls.admin_urls")),
+    path("api/users/", include("apps.users.urls", namespace="users")),
     path("api/vi/accounts", include("apps.users.urls", namespace="users")),
 ]
